把业务服务改成接口,方便扩展其它业务接口

+++ internal/bu_service/bu_system_service_api.go
@@ -2,6 +2,7 @@ package bu_service
 
 import (
 	log "github.com/sirupsen/logrus"
+	"jt809_server/config"
 	"jt809_server/models"
 	"jt809_server/util"
 	"sync"
@@ -12,7 +13,7 @@ import (
 // VsBuServiceApi 业务系统服务对外接口,单例
 
 type VsBuSystemServiceApi struct {
-	ctx *buSystemServiceContext
+	ctx buSystemBase
 }
 
 var api *VsBuSystemServiceApi
@@ -26,21 +27,45 @@ func MsgServiceIns() *VsBuSystemServiceApi {
 	return api
 }
 
+// vsBuType 业务系统类型,可以通过配置文件来更改
+type vsBuType int
+
+const (
+	VS_BU_TYPE_SELF_V4  vsBuType = 0 //对接自己V4平台
+	VS_BU_TYPE_OTHER_V1 vsBuType = 1 //其它平台
+)
+
 func (_self *VsBuSystemServiceApi) InitApi(notifyFrameWork *util.IotMsgNotifyFramework) {
-	_self.ctx = &buSystemServiceContext{}
-	_self.ctx.StartService("buSystemService", 50000, notifyFrameWork, func() {
-		log.Info("buSystemService start finish")
-		_self.ctx.initService()
-	})
+	buType := config.GetBuType()
+	log.Info("init bu type:", buType)
+	if VS_BU_TYPE_OTHER_V1 == vsBuType(buType) {
+
+	} else { //默认 VS_BU_TYPE_SELF_V4
+		selfBuCtx := &buSystemServiceContext{}
+		selfBuCtx.StartService("buSystemService", 50000, notifyFrameWork, func() {
+			log.Info("buSystemService start finish")
+			selfBuCtx.initService()
+		})
+		_self.ctx = selfBuCtx
+	}
+
 }
 
 // AddBuUserInfo 添加用户信息
 func (_self *VsBuSystemServiceApi) AddBuUserInfo(userName, userPwdMd5 string, userKey int64) {
+	if _self.ctx == nil {
+		log.Error("AddBuUserInfo fail,not init bu ctx")
+		return
+	}
 	_self.ctx.addBuUserInfo(userName, userPwdMd5, userKey)
 }
 
 // DeleteBuUserInfo 删除用户信息
 func (_self *VsBuSystemServiceApi) DeleteBuUserInfo(userKey int64) {
+	if _self.ctx == nil {
+		log.Error("DeleteBuUserInfo fail,not init bu ctx")
+		return
+	}
 	_self.ctx.deleteBuUserInfo(userKey)
 }
 
@@ -48,11 +73,19 @@ func (_self *VsBuSystemServiceApi) DeleteBuUserInfo(userKey int64) {
 
 // SendDevTakePhotoReq 请求抓拍
 func (_self *VsBuSystemServiceApi) SendDevTakePhotoReq(userKey int64, reqObject *models.VsAppTakePhotoReq) bool {
+	if _self.ctx == nil {
+		log.Error("SendDevTakePhotoReq fail,not init bu ctx")
+		return false
+	}
 	return _self.ctx.sendDevTakePhotoReq(userKey, reqObject)
 }
 
 // AsyncSendPlatformPostQueryReq 上级平台查岗
 func (_self *VsBuSystemServiceApi) AsyncSendPlatformPostQueryReq(userKey int64, reqObject *models.VsAppMsgPlatformPostQueryReq) {
+	if _self.ctx == nil {
+		log.Error("AsyncSendPlatformPostQueryReq fail,not init bu ctx")
+		return
+	}
 	go func() {
 		_self.ctx.sendPlatformPostQueryReq(userKey, reqObject)
 	}()
@@ -60,6 +93,10 @@ func (_self *VsBuSystemServiceApi) AsyncSendPlatformPostQueryReq(userKey int64, 
 
 // AsyncSendPlatformMsgTextInfoReq 上级平台报文下发
 func (_self *VsBuSystemServiceApi) AsyncSendPlatformMsgTextInfoReq(userKey int64, reqObject *models.VsAppMsgPlatformTextReq) {
+	if _self.ctx == nil {
+		log.Error("AsyncSendPlatformMsgTextInfoReq fail,not init bu ctx")
+		return
+	}
 	go func() {
 		_self.ctx.sendPlatformMsgTextInfoReq(userKey, reqObject)
 	}()
@@ -68,6 +105,10 @@ func (_self *VsBuSystemServiceApi) AsyncSendPlatformMsgTextInfoReq(userKey int64
 
 // AsyncSendPlatformWarnUrgeTodoReq 报警督办请求
 func (_self *VsBuSystemServiceApi) AsyncSendPlatformWarnUrgeTodoReq(userKey int64, reqObject *models.VsAppMsgWarnUrgeTodoReq) {
+	if _self.ctx == nil {
+		log.Error("AsyncSendPlatformWarnUrgeTodoReq fail,not init bu ctx")
+		return
+	}
 	go func() {
 		_self.ctx.sendPlatformWarnUrgeTodoReq(userKey, reqObject)
 	}()
